Add files via upload
Python provides several types of loops to execute a block of code multiple times. The most common ones are for loops and while loops. Here's a brief overview of each:

for Loop
A for loop is used for iterating over a sequence (such as a list, tuple, dictionary, set, or string). This is the most common use case for for loops.
while Loop
A while loop continues to execute a block of code as long as the specified condition is True.